requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-jose
passlib[bcrypt]
pydantic
python-dotenv
werkzeug
